Extract CanvasProjectAccessControls from CanvasPage
